requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

langgraph==0.0.30
langchain-core==0.1.52

supabase>=2.27.2

openai==1.12.0

python-dotenv==1.0.1
pydantic>=2.11.7,<3.0.0

tenacity==8.2.3
# httpx and gotrue are managed by supabase - no need to pin them

pytest==8.1.1

# PDF processing for RAG
pypdf==4.0.1